requirements.txt
fastapi
uvicorn[standard]
scikit-learn==1.8.0
mlflow==3.8.1
python-dotenv==1.2.1
prometheus-client==0.23.1
joblib==1.5.3
pytest==9.0.2
requests==2.32.5
pandas==2.3.3
matplotlib==3.10.0
seaborn==0.13.2
flake8==7.1.1
black==24.10.0
pytest-cov==6.0.0
opencensus-ext-azure==1.1.13
